Plot an object of class mixture

provides different diagnostic plots for an aggregation procedure.

# S3 method for mixture
plot(x, pause = FALSE, col = NULL, ...)

Arguments

x
an object of class mixture. If awake is provided (i.e., some experts are unactive), their residuals and cumulative losses are computed by using the predictions of the mixture.

pause
if set to TRUE (default) displays the plots separately, otherwise on a single page

col
the color to use to represent each experts, if set to NULL (default) use RRColorBrewer::brewer.pal(...,"Spectral"

...
additional plotting parameters

Value

plots representing: plot of weights of each expert in function of time, boxplots of these weights, cumulative loss \(L_T=\sum_{t=1}^T l_{i,t}\) of each expert in function of time, cumulative residuals \(\sum_{t=1}^T (y_t-f_{i,t})\) of each expert's forecast in function of time, average loss suffered by the experts and the contribution of each expert to the aggregation \(p_{i,t}f_{i,t}\) in function of time.

See also

Author

Pierre Gaillard <pierre@gaillard.me>

Yannig Goude <yannig.goude@edf.fr>